update for cors area, update functionality of posting information from the application
