+++ package.json
@@ -5,11 +5,13 @@
   "private": true,
   "dependencies": {
     "axios": "^0.18.0",
+    "chart.js": "^2.8.0",
     "firebase": "^5.10.1",
     "gh-pages": "^2.0.1",
     "google-map-react": "^1.1.4",
     "hover.css": "^2.3.2",
     "react": "^16.8.6",
+    "react-chartjs-2": "^2.7.6",
     "react-dom": "^16.8.6",
     "react-materialize": "^3.2.0",
     "react-router-dom": "^5.0.0",
