+++ src/App.js
@@ -14,6 +14,7 @@ import AdminDeniedApp from "./pages/admin/deniedApplications";
 import AdminApprovedApp from "./pages/admin/approvedApplications";
 import User from "./pages/user";
 import Trucker from "./pages/trucker";
+import TruckerApplication from "./pages/trucker/Food Truck Application";
 
 // import AppTwo from './pages/Signin/AppTwo';
 import NavigationAuth from './pages/Signin/NavigationAuth';
@@ -55,6 +56,7 @@ function App() {
         <Route exact path="/admin/view/applications/approved" component={AdminApprovedApp} />
         <Route exact path="/user/dashboard" component={User} />
         <Route exact path="/trucker/dashboard" component={Trucker} />
+        <Route exact path="/trucker/application" component={TruckerApplication} />
         <Route exact path="*" component={Splash} />
       </Switch>
     </Router>
