+++ src/pages/admin/deniedApplications/index.js
@@ -1,20 +1,21 @@
 import React, { Component } from "react";
 import Nav from "../../../components/admin/navToHome";
 import Container from "../../../components/admin/container";
-import API from "../../../utils/API"
+import Chart from "../../../components/admin/chart";
+import {Bar, Line, Pie} from "react-chartjs-2";
 
 class Admin extends Component {
 
     render() {
       return (
-        <div className="brickBackground">
+        <div className="">
           <Nav 
           currentPage="Denied Applications"
           />
-          <Container>
-          <h1>Whats UP WHYYYY CAIN hello hello hello hello</h1>
+         
+          <Chart />
  
-          </Container>
+         
         </div>)
     }
